"Test Task Uploaded 12:30"

+++ test/register.js
@@ -25,14 +25,22 @@ else
         let name = document.getElementById('name').value;
         let email = document.getElementById('email').value;
         let password = document.getElementById('password').value;
+        let confirmpassword = document.getElementById('confirmpassword').value;
         let city = document.getElementById('city').value;
         let state = document.getElementById('state').value;
 
-        if(name !== '' && email !== '' && password !== '' && city !== '' && state !== '')
+        if(name !== '' && email !== '' && password !== '' && confirmpassword !== '' && city !== '' && state !== '')
         {
-            adminArray.push(new Admin(name,email,password,city,state));
-            localStorage.setItem('admin',JSON.stringify(adminArray));
-            alert('Admin Register Successfully !!');
+            if(password === confirmpassword)
+            {
+                adminArray.push(new Admin(name,email,password,city,state));
+                localStorage.setItem('admin',JSON.stringify(adminArray));
+                alert('Admin Register Successfully !!');
+            }
+            else
+            {
+                alert("Password Not Match !!");
+            }
         }
         else
         {
